Functionality to add Chit Schema #28
* Fixed the Helper Util bug
* Chit scheme saveing properly, but all fields are not yet added in WPF UI.
* Added my device id in seeding in debug mode.

+++ src/Api.Database.Entity/User/User.cs
@@ -9,7 +9,7 @@ namespace Api.Database.Entity.User
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Key, Column(Order = 0)]
-        public UInt64 Id { get; set; }
+        public int Id { get; set; }
         public string UserId { get; set; }
         public string Password { get; set; }
         public string ConfirmPassword { get; set; }
